+++ src/shapes/Path.js
@@ -125,11 +125,30 @@
         height: Math.round(maxY - minY)
       };
     },
-    getLength: function () {
+    /**
+     * Return length of the path.
+     * @method
+     * @memberof Konva.Path.prototype
+     * @returns {Number} length
+     * @example
+     * var length = path.getLength();
+     */
+    getLength: function() {
       return this.pathLength;
     },
-    getPointAtLength: function (length) {
-      var point, i = 0, ii = this.dataArray.length;
+    /**
+     * Get point on path at specific length of the path
+     * @method
+     * @memberof Konva.Path.prototype
+     * @param {Number} length length
+     * @returns {Object} point {x,y} point
+     * @example
+     * var point = path.getPointAtLength(10);
+     */
+    getPointAtLength: function(length) {
+      var point,
+        i = 0,
+        ii = this.dataArray.length;
 
       if (!ii) {
         return null;
@@ -160,19 +179,48 @@
       var p = cp.points;
       switch (cp.command) {
         case 'L':
-          return Konva.Path.getPointOnLine(length, cp.start.x, cp.start.y, p[0], p[1]);
+          return Konva.Path.getPointOnLine(
+            length,
+            cp.start.x,
+            cp.start.y,
+            p[0],
+            p[1]
+          );
         case 'C':
-          return Konva.Path.getPointOnCubicBezier(length / cp.pathLength, cp.start.x, cp.start.y, p[0], p[1], p[2], p[3], p[4], p[5]);
+          return Konva.Path.getPointOnCubicBezier(
+            length / cp.pathLength,
+            cp.start.x,
+            cp.start.y,
+            p[0],
+            p[1],
+            p[2],
+            p[3],
+            p[4],
+            p[5]
+          );
         case 'Q':
-          return Konva.Path.getPointOnQuadraticBezier(length / cp.pathLength, cp.start.x, cp.start.y, p[0], p[1], p[2], p[3]);
+          return Konva.Path.getPointOnQuadraticBezier(
+            length / cp.pathLength,
+            cp.start.x,
+            cp.start.y,
+            p[0],
+            p[1],
+            p[2],
+            p[3]
+          );
         case 'A':
-          var cx = p[0], cy = p[1], rx = p[2], ry = p[3], theta = p[4], dTheta = p[5], psi = p[6];
+          var cx = p[0],
+            cy = p[1],
+            rx = p[2],
+            ry = p[3],
+            theta = p[4],
+            dTheta = p[5],
+            psi = p[6];
           theta += dTheta * length / cp.pathLength;
           return Konva.Path.getPointOnEllipticalArc(cx, cy, rx, ry, theta, psi);
       }
 
       return null;
-
     }
   };
   Konva.Util.extend(Konva.Path, Konva.Shape);